Requirement.java and TestDefinition.java has moved into hu.blackbelt.judo.test package. Some test cases in the PrimitivesTest.java has been annotated with Requirment.
