+++ judo-runtime-core-jsl-itest/src/test/java/hu/blackbelt/judo/runtime/core/jsl/PrimitivesTest.java
@@ -27,6 +27,7 @@ import hu.blackbelt.judo.dispatcher.api.FileType;
 import hu.blackbelt.judo.runtime.core.exception.ValidationException;
 import hu.blackbelt.judo.runtime.core.jsl.itest.primitives.guice.primitives.PrimitivesDaoModules;
 import hu.blackbelt.judo.runtime.core.jsl.itest.primitives.sdk.primitives.primitives.*;
+import hu.blackbelt.judo.test.Requirement;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 
@@ -79,7 +80,12 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
-    @TestDefinition(testDefs={"TCD-008"})
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-003",
+            "REQ-ENT-002"
+    })
     public void testPlainOptionalEntityCreationWithoutValues() {
         MyEntityWithOptionalFields myEntityWithOptionalFields = myEntityWithOptionalFieldsDao.create(MyEntityWithOptionalFields.builder().build());
 
@@ -99,7 +105,19 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
-    @TestDefinition(testDefs={"TCD-009"})
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-003",
+            "REQ-TYPE-004",
+            "REQ-TYPE-005",
+            "REQ-TYPE-006",
+            "REQ-TYPE-007",
+            "REQ-TYPE-008",
+            "REQ-TYPE-009",
+            "REQ-TYPE-010",
+            "REQ-ENT-002"
+    })
     public void testPlainOptionalEntityCreationWithValues() {
         MyEntityWithOptionalFields myEntityWithOptionalFields = myEntityWithOptionalFieldsDao.create(MyEntityWithOptionalFields.builder()
                         .withIntegerAttr(1)
@@ -131,6 +149,12 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test()
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-003",
+            "REQ-ENT-002"
+    })
     public void testMissingRequiredFieldsThrowExceptions() {
         ValidationException thrown = assertThrows(
             ValidationException.class,
@@ -173,6 +197,19 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-003",
+            "REQ-TYPE-004",
+            "REQ-TYPE-005",
+            "REQ-TYPE-006",
+            "REQ-TYPE-007",
+            "REQ-TYPE-009",
+            "REQ-TYPE-010",
+            "REQ-ENT-002",
+            "REQ-ENT-003"
+    })
     public void testIdentifierFieldsAreUnique() {
         LocalDate now = LocalDate.now();
         EntityWithIdentifiers ent1 = entityWithIdentifiersDao.create(EntityWithIdentifiers.builder()
@@ -236,6 +273,16 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-003",
+            "REQ-TYPE-004",
+            "REQ-TYPE-005",
+            "REQ-TYPE-006",
+            "REQ-TYPE-007",
+            "REQ-ENT-002"
+    })
     public void testFieldsAreNonUnique() {
         LocalDate now = LocalDate.now();
         MyEntityWithOptionalFields ent1 = myEntityWithOptionalFieldsDao.create(MyEntityWithOptionalFields.builder()
@@ -260,6 +307,19 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-003",
+            "REQ-TYPE-004",
+            "REQ-TYPE-005",
+            "REQ-TYPE-006",
+            "REQ-TYPE-007",
+            "REQ-TYPE-008",
+            "REQ-TYPE-009",
+            "REQ-TYPE-010",
+            "REQ-ENT-002"
+    })
     public void testUpdateOptionalEntityAfterCreation() {
         MyEntityWithOptionalFields myEntityWithOptionalFields = myEntityWithOptionalFieldsDao.create(MyEntityWithOptionalFields.builder().build());
 
@@ -304,6 +364,18 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-003",
+            "REQ-TYPE-004",
+            "REQ-TYPE-005",
+            "REQ-TYPE-006",
+            "REQ-TYPE-007",
+            "REQ-TYPE-008",
+            "REQ-TYPE-009",
+            "REQ-ENT-002"
+    })
     public void testEntityCreationWithPrimitiveDefaults() {
         EntityWithPrimitiveDefaults entityWithDefaults = entityWithPrimitiveDefaultsDao.create(EntityWithPrimitiveDefaults.builder().build());
 
@@ -326,6 +398,18 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-003",
+            "REQ-TYPE-004",
+            "REQ-TYPE-005",
+            "REQ-TYPE-006",
+            "REQ-TYPE-007",
+            "REQ-TYPE-008",
+            "REQ-TYPE-009",
+            "REQ-ENT-002"
+    })
     public void testEntityCreationRequiredWithPrimitiveDefaults() {
         EntityRequiredWithPrimitiveDefaults entityRequiredWithDefaults = entityRequiredWithPrimitiveDefaultsDao.create(EntityRequiredWithPrimitiveDefaults.builder().build());
 
@@ -347,6 +431,22 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-003",
+            "REQ-TYPE-004",
+            "REQ-TYPE-005",
+            "REQ-TYPE-006",
+            "REQ-TYPE-007",
+            "REQ-TYPE-008",
+            "REQ-TYPE-009",
+            "REQ-ENT-002",
+            "REQ-EXPR-014",
+            "REQ-EXPR-015",
+            "REQ-EXPR-002",
+            "REQ-EXPR-009",
+            "REQ-EXPR-017"
+    })
     public void testEntityCreationWithPrimitiveDefaultExpressions() {
         EntityWithPrimitiveDefaultExpressions entityWithDefaultExpressions = entityWithPrimitiveDefaultExpressionsDao.create(EntityWithPrimitiveDefaultExpressions.builder().build());
 
@@ -366,6 +466,12 @@ public class PrimitivesTest extends AbstractJslTest {
     }
 
     @Test
+    @Requirement(reqs = {
+            "REQ-TYPE-001",
+            "REQ-TYPE-003",
+            "REQ-TYPE-004",
+            "REQ-ENT-002"
+    })
     public void testRegexValidatorFailsForInvalidInput() {
         ValidationException thrown = assertThrows(
                 ValidationException.class,
