+++ judo-runtime-core-jsl-itest/src/test/java/hu/blackbelt/judo/test/TestDefinition.java
@@ -1,4 +1,4 @@
-package hu.blackbelt.judo.runtime.core.jsl;
+package hu.blackbelt.judo.test;
 
 /*-
  * #%L
